Adding institution all, throwing unused endpoints
