+++ src/main/java/com/plaid/client/DefaultPlaidPublicClient.java
@@ -4,12 +4,7 @@ import com.plaid.client.exception.PlaidClientsideException;
 import com.plaid.client.http.HttpDelegate;
 import com.plaid.client.http.HttpResponseWrapper;
 import com.plaid.client.http.PlaidHttpRequest;
-import com.plaid.client.request.MappingOptions;
-import com.plaid.client.response.CategoriesResponse;
-import com.plaid.client.response.Category;
-import com.plaid.client.response.Institution;
 import com.plaid.client.response.InstitutionsResponse;
-import com.plaid.client.response.LongTailInstitutionsResponse;
 import org.apache.commons.lang.StringUtils;
 
 public class DefaultPlaidPublicClient implements PlaidPublicClient {
@@ -20,30 +15,11 @@ public class DefaultPlaidPublicClient implements PlaidPublicClient {
     private String secret;
 
     @Override
-    public Object getEntity(String entityId) {
-        throw new UnsupportedOperationException("Not implemented yet");
-    }
-
-    @Override
-    public Object getInstitution(String institutionId) {
-        throw new UnsupportedOperationException("Not implemented yet");
-    }
-
-    @Override
-    public InstitutionsResponse getAllInstitutions() {
-        PlaidHttpRequest request = new PlaidHttpRequest("/institutions");
-        HttpResponseWrapper<Institution[]> response = httpDelegate.doGet(request, Institution[].class);
-        return new InstitutionsResponse(response.getResponseBody());
-    }
-
-    @Override
-    public LongTailInstitutionsResponse getAllLongTailInstitutions(Integer offset, Integer count) {
-        PlaidHttpRequest request = new PlaidHttpRequest("/institutions/longtail");
-
+    public InstitutionsResponse getAllInstitutions(Integer offset, Integer count) {
+        PlaidHttpRequest request = new PlaidHttpRequest("/institutions/all");
         if(StringUtils.isEmpty(clientId) || StringUtils.isEmpty(secret)) {
             throw new PlaidClientsideException("ClientId and Secret is required");
         }
-
         request.addParameter("client_id", clientId);
         request.addParameter("secret", secret);
 
@@ -54,28 +30,10 @@ public class DefaultPlaidPublicClient implements PlaidPublicClient {
         if(offset != null) {
             request.addParameter("offset", offset.toString());
         }
-
-        HttpResponseWrapper<LongTailInstitutionsResponse> response = httpDelegate.doPost(request, LongTailInstitutionsResponse.class);
+        HttpResponseWrapper<InstitutionsResponse> response = httpDelegate.doPost(request, InstitutionsResponse.class);
         return response.getResponseBody();
     }
 
-    @Override
-    public CategoriesResponse getAllCategories() {
-        PlaidHttpRequest request = new PlaidHttpRequest("/categories");
-        HttpResponseWrapper<Category[]> response = httpDelegate.doGet(request, Category[].class);
-        return new CategoriesResponse(response.getResponseBody());
-    }
-
-    @Override
-    public Object getCategory(String categoryId) {
-        throw new UnsupportedOperationException("Not implemented yet");
-    }
-
-    @Override
-    public Object getCategoriesByMapping(String mapping, MappingOptions options) {
-        throw new UnsupportedOperationException("Not implemented yet");
-    }
-    
     @Override
     public HttpDelegate getHttpDelegate() {
     	return httpDelegate;
@@ -111,7 +69,6 @@ public class DefaultPlaidPublicClient implements PlaidPublicClient {
 
             return client;
         }
-
     }
 
 }
