+++ src/main/java/com/plaid/client/PlaidPublicClient.java
@@ -1,26 +1,11 @@
 package com.plaid.client;
 
 import com.plaid.client.http.HttpDelegate;
-import com.plaid.client.request.MappingOptions;
-import com.plaid.client.response.CategoriesResponse;
 import com.plaid.client.response.InstitutionsResponse;
-import com.plaid.client.response.LongTailInstitutionsResponse;
 
 public interface PlaidPublicClient {
-
-    Object getEntity(String entityId);
     
-    Object getInstitution(String institutionId);
-    
-    InstitutionsResponse getAllInstitutions();
-
-    LongTailInstitutionsResponse getAllLongTailInstitutions(Integer offset, Integer count);
-    
-    CategoriesResponse getAllCategories();
-    
-    Object getCategory(String categoryId);
-    
-    Object getCategoriesByMapping(String mapping, MappingOptions options);
+    InstitutionsResponse getAllInstitutions(Integer offset, Integer count);
     
     HttpDelegate getHttpDelegate();
 }
