+++ src/main/java/com/plaid/client/response/InstitutionsResponse.java
@@ -1,17 +1,20 @@
 package com.plaid.client.response;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown=true)
 public class InstitutionsResponse {
 
-    private Institution[] institutions;
+    private Integer totalCount;
+    private Institution[] results;
 
-    public InstitutionsResponse(Institution[] institutions) {
-        this.institutions = institutions;
+    @JsonProperty("total_count")
+    public Integer getTotalCount() {
+        return totalCount;
     }
 
-    public Institution[] getInstitutions() {
-        return institutions;
+    public Institution[] getResults() {
+        return results;
     }
 }
