--- src/main/java/com/plaid/client/response/LongTailInstitution.java
@@ -1,26 +1,0 @@
-package com.plaid.client.response;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@JsonIgnoreProperties(ignoreUnknown=true)
-public class LongTailInstitution extends Institution {
-
-    private String currencyCode;
-    private String url;
-
-    public String getCurrencyCode() {
-        return currencyCode;
-    }
-
-    public void setCurrencyCode(String currencyCode) {
-        this.currencyCode = currencyCode;
-    }
-
-    public String getUrl() {
-        return url;
-    }
-
-    public void setUrl(String url) {
-        this.url = url;
-    }
-}
