+++ src/test/java/com/plaid/client/PlaidPublicClientTest.java
@@ -1,14 +1,11 @@
 package com.plaid.client;
 
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
 
 import java.util.HashMap;
 import java.util.Map;
 
 import com.plaid.client.exception.PlaidClientsideException;
-import com.plaid.client.response.LongTailInstitutionsResponse;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
 import org.junit.Before;
@@ -16,24 +13,18 @@ import org.junit.Test;
 
 import com.plaid.client.http.ApacheHttpClientHttpDelegate;
 import com.plaid.client.http.HttpDelegate;
-import com.plaid.client.response.CategoriesResponse;
 import com.plaid.client.response.Institution;
 import com.plaid.client.response.InstitutionsResponse;
 
 public class PlaidPublicClientTest {
     
     private CloseableHttpClient httpClient;
-    private HttpDelegate httpDelegate;
     private PlaidPublicClient plaidPublicClientWithoutCredentials;
     private PlaidPublicClient plaidPublicClientWithCredentials;
-    
-    //@Rule
-    //public WireMockRule wireMockRule = new WireMockRule(8089);
 
     @Before
     public  void setup() {
         httpClient = HttpClients.custom().disableContentCompression().build();
-        //httpDelegate = new ApacheHttpClientHttpDelegate("http://localhost:8089", httpClient);
         HttpDelegate httpDelegate = new ApacheHttpClientHttpDelegate("https://tartan.plaid.com", httpClient);
         plaidPublicClientWithoutCredentials = new DefaultPlaidPublicClient.Builder()
                 .withHttpDelegate(httpDelegate)
@@ -45,18 +36,12 @@ public class PlaidPublicClientTest {
                 .withSecret("test_secret")
                 .build();
     }
-    
-    @Test
-    public void testGetAllCategories() {
-        CategoriesResponse categoriesResponse = plaidPublicClientWithoutCredentials.getAllCategories();
-        assertTrue(categoriesResponse.getCategories().size() > 0);
-    }
 
     @Test
     public void testGetAllInstitutions() {
-        InstitutionsResponse instResponse = plaidPublicClientWithoutCredentials.getAllInstitutions();
+        InstitutionsResponse instResponse = plaidPublicClientWithCredentials.getAllInstitutions(0, 10);
         assertNotNull(instResponse);
-        Institution[] institutions = instResponse.getInstitutions();
+        Institution[] institutions = instResponse.getResults();
         Map<String, Institution> map = new HashMap<>();
         for (Institution institution : institutions) {
             map.put(institution.getName(), institution);
@@ -66,13 +51,7 @@ public class PlaidPublicClientTest {
     }
 
     @Test(expected = PlaidClientsideException.class)
-    public void testGetAllLongTailInstitutionsRequireCredentials() throws Exception {
-        plaidPublicClientWithoutCredentials.getAllLongTailInstitutions(0, 10);
-    }
-
-    @Test
-    public void testGetAllLongTailInstitutions() throws Exception {
-        LongTailInstitutionsResponse response = plaidPublicClientWithCredentials.getAllLongTailInstitutions(0, 10);
-        assertEquals(10, response.getResults().length);
+    public void testGetAllInstitutionsRequireCredentials() throws Exception {
+        plaidPublicClientWithoutCredentials.getAllInstitutions(0, 10);
     }
 }
